device_util: Make device in dev_find_slot_pnp u16.
LDN is 8-bit but coreboot squeezes unrelated info: VLDN in this field.
Increase to 16-bit to handle this.

Reviewed-on: http://review.coreboot.org/5165
Tested-by: build bot (Jenkins)

+++ src/include/device/device.h
@@ -169,7 +169,7 @@ device_t dev_find_device (u16 vendor, u16 device, device_t from);
 device_t dev_find_class (unsigned int class, device_t from);
 device_t dev_find_slot (unsigned int bus, unsigned int devfn);
 device_t dev_find_slot_on_smbus (unsigned int bus, unsigned int addr);
-device_t dev_find_slot_pnp(u16 port, u8 device);
+device_t dev_find_slot_pnp(u16 port, u16 device);
 device_t dev_find_lapic(unsigned apic_id);
 int dev_count_cpu(void);
